ARCHITECTURAL/ArchiGPT_0.2/datasets/initial/2_Marketing_Analyzer/Marketing-Analyzer-master/Inference_Server/requirements.txt
Flask
Flask-RESTful
flask_cors
pandas
sklearn
pymongo[srv]
python-dotenv